wip: create ci/cd scaffold
