+++ packages/cli-plugin-scaffold-ci/src/githubActions/workflows/.github/workflows/webinyDevPr.yml
@@ -1,0 +1,85 @@
+name: Dev Branch - Pull Request
+
+on:
+  pull_request:
+    branches: [dev]
+
+env:
+  NODE_OPTIONS: --max_old_space_size=4096
+
+jobs:
+  cache-dependencies:
+    name: Cache dependencies
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/setup-node@v2
+        with:
+          node-version: 12
+
+      - uses: actions/checkout@v2
+
+      - uses: actions/cache@v2
+        id: yarn-cache
+        with:
+          path: .yarn/cache
+          key: yarn-${{ runner.os }}-${{ hashFiles('**/yarn.lock') }}
+
+      - name: Install dependencies
+        if: steps.yarn-cache.outputs.cache-hit != 'true'
+        run: yarn --immutable
+
+  code-analysis:
+    needs: cache-dependencies
+    name: Static code analysis
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/setup-node@v2
+        with:
+          node-version: 12
+
+      - uses: actions/checkout@v2
+
+      - uses: actions/cache@v2
+        with:
+          path: .yarn/cache
+          key: yarn-${{ runner.os }}-${{ hashFiles('**/yarn.lock') }}
+
+      - name: Install dependencies
+        run: yarn --immutable
+
+      # - name: Check code formatting
+      #  run: yarn prettier:check
+
+      - name: ESLint
+        run: yarn eslint
+
+  jest-tests:
+    needs: cache-dependencies
+    name: Jest tests
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest]
+        node: [12, 14]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - uses: actions/setup-node@v2
+        with:
+          node-version: ${{ matrix.node }}
+
+      - uses: actions/checkout@v2
+
+      - uses: actions/cache@v2
+        with:
+          path: .yarn/cache
+          key: yarn-${{ runner.os }}-${{ hashFiles('**/yarn.lock') }}
+
+      - name: Install dependencies
+        run: yarn --immutable
+
+      - name: Build packages
+        run: yarn webiny workspaces run build --folder packages
+
+      - name: Run tests
+        run: yarn test
+
